\name{oclPlatforms}
\alias{oclPlatforms}
\title{
Retrieve available OpenCL platforms.
}
\description{
\code{oclPlatforms} retrieves all available OpenCL platforms.
}
\usage{
oclPlatforms()
}
%\details{
%}
\value{
  List of available OpenCL platforms.
}
%\references{
%}
\author{
Simon Urbanek
}
\seealso{
\code{\link{oclDevices}}
}
\examples{
print(oclPlatforms())
}
\keyword{interface}
